Rename ServerInfoLedger to ServerInfoValidatedLedger
