+++ xrpl4j-model/src/test/java/org/xrpl/xrpl4j/model/client/serverinfo/JobTypeTest.java
@@ -6,6 +6,9 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import com.google.common.primitives.UnsignedInteger;
 import org.junit.jupiter.api.Test;
 
+/**
+ * Unit tests for {@link JobType}.
+ */
 public class JobTypeTest {
 
   @Test
